requirements.txt
streamlit
google-generativeai
pytrends
python-dotenv